connect view profile
