+++ Client/HirePro/lib/Models/Task.dart
@@ -1,44 +1,44 @@
-import 'dart:ffi';
-import 'dart:io';
-
-class TaskDetails {
-  String id;
-  String customerID;
-  String serviceID;
-  String description;
-  String acceptedTime;
-  String latitude;
-  String longitude;
-  String status;
-  String estMin;
-  String estMax;
-  File? image;
-  File? verifyImages;
-  Int points;
-  String intro;
-
-  TaskDetails({
-    // required this.id,
-    // required this.name,
-    // required this.email,
-    // required this.contact,
-    // required this.nic,
-    // this.image,
-    // required this.verifyImages,
-    // required this.points,
-    // required this.intro,
-  });
-
-  factory TaskDetails.fromJson(Map<String, dynamic> json) {
-    return TaskDetails(
-        // id: json['id'],
-        // name: json['name'],
-        // email: json['email'],
-        // contact: json['contact'],
-        // nic: json['nic'],
-        // verifyImages: json['verify_image'],
-        // points: json['points'],
-        // intro: json['intro']
-    );
-  }
-}
+// import 'dart:ffi';
+// import 'dart:io';
+//
+// class TaskDetails {
+//   String id;
+//   String customerID;
+//   String serviceID;
+//   String description;
+//   String acceptedTime;
+//   String latitude;
+//   String longitude;
+//   String status;
+//   String estMin;
+//   String estMax;
+//   File? image;
+//   File? verifyImages;
+//   Int points;
+//   String intro;
+//
+//   TaskDetails({
+//     // required this.id,
+//     // required this.name,
+//     // required this.email,
+//     // required this.contact,
+//     // required this.nic,
+//     // this.image,
+//     // required this.verifyImages,
+//     // required this.points,
+//     // required this.intro,
+//   });
+//
+//   factory TaskDetails.fromJson(Map<String, dynamic> json) {
+//     return TaskDetails(
+//         // id: json['id'],
+//         // name: json['name'],
+//         // email: json['email'],
+//         // contact: json['contact'],
+//         // nic: json['nic'],
+//         // verifyImages: json['verify_image'],
+//         // points: json['points'],
+//         // intro: json['intro']
+//     );
+//   }
+// }
