+++ Server/request.rest
@@ -45,4 +45,4 @@ Authorization: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiNiIsInVzZXJf
 
 ###
 GET http://localhost:5001/getUser
-Authorization: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiNiIsInVzZXJfbmFtZSI6IlRoYXJ1c2hrYSBNYWR1c2FyYSIsInVzZXJfZW1haWwiOiJtYWR1c2FyYUBnbWFpbC5jb20iLCJpYXQiOjE2OTIwNjUwMDYsImV4cCI6MTY5MzI3NDYwNn0.fg_6UuLqd5HKkdRP0wgi-voDNFUUDW_AohIL8Qjr8vk
+Authorization: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiMjMiLCJ1c2VyX25hbWUiOiJzYWNoaW5pIG11dGh1Z2FsYSIsInVzZXJfZW1haWwiOiJzQGdtYWlsLmNvbSIsImlhdCI6MTY5NDMyNzM0NCwiZXhwIjoxNjk1NTM2OTQ0fQ.slDaDoq1lHad4ysbEkea2kBYnELTJho60Khd_KZGz14
